Move more files to WebUI
Now everything with Luc's copyright is in WebUI/
