+++ Grbl_Esp32/src/System.h
@@ -1,7 +1,7 @@
 #pragma once
 
 /*
-  system.h - Header for system level commands and real-time processes
+  System.h - Header for system level commands and real-time processes
   Part of Grbl
   Copyright (c) 2014-2016 Sungeun K. Jeon for Gnea Research LLC
 
@@ -22,7 +22,6 @@
 
 #include "Grbl.h"
 #include "TDef.h"
-#include "Commands.h"
 
 // Define global system variables
 typedef struct {
