+++ Grbl_Esp32/src/WebUI/BTConfig.cpp
@@ -18,7 +18,7 @@
   Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 */
 
-#include "Grbl.h"
+#include "../Grbl.h"
 
 #ifdef ENABLE_BLUETOOTH
 #    include <BluetoothSerial.h>
